schedule auto peer review properly on import
unfortunately, when assignments are imported and dates are specified to
be shifted, the assignment is first saved with the old dates, and then
the date shift is applied and the assignment is saved again.  on the
first save, the auto peer review job is scheduled at the old time, and
then when the assignment is re-saved, because the job is a singleton and
the new time is later than the old time, the job is not re-scheduled at
the proper time. this change just adds a flag to prevent the first job
from being scheduled.

fixes CNVS-18243

test plan:
- create a course with start and end dates
- create an assignment with auto peer review scheduled in the future
- go to /jobs, to the future tab, and you should see the job scheduled
- create a new course with start and end dates 1 week later than the old
  course
- go to import content, copy from the old course, shift dates
- shift the dates from the old course start/end to the new course
  start/end
- import the assignment
- go to /jobs, to the future tab
- there should be a second scheduled job, set to run at the
  appropriately shifted time

Change-Id: I90631cfbff1acaa2e3fb009db0d12dd5cdd5694e
Reviewed-on: https://gerrit.instructure.com/50739
Tested-by: Jenkins
Reviewed-by: John Corrigan <jcorrigan@instructure.com>
QA-Review: Adam Stone <astone@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>