make Download all submissions a modal instead of a pseudo-modal
fixes CNVS-19217

When the Download all submissions button is pressed it behaved
similarly to a modal, but isn't one because it lets focus outside
of the modal. This change uses the jQuery UI modal to handle most
of the heavy lifting, and removes code which was used to create
this pseudo-modal behavior.

test plan:

1. In the individual gradebook view, select an assignment
    under the Content Selection menu.
2. Click the "download all submissions" button.
3. Try tabbing. No amount of tabbing or shift-tabbing should
    allow your focus to be outside of the modal.

Change-Id: I03453293e0ee20d3f43f0a1f4cb5c58ef980924c
Reviewed-on: https://gerrit.instructure.com/51265
Tested-by: Jenkins
Reviewed-by: Jonathan Featherstone <jfeatherstone@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Strand McCutchen <smccutchen@instructure.com>