makes Google Doc tab more accessible
fixes CNVS-19813

When a user activates a tab either by clicking or tabbing
with arrow keys on their keyboard, Google Docs loads a list
of their documents.

When a user selects a document either by clicking or using
the spacebar, the document is selected, which is indicated
with a grey background.

When a user opens a folder either by clicking or using
the spacebar, the folder expands.

Test plan:

On an account with the Google Drive/Doc plugins setup, as
a student with Google Drive set up in their user settings.
For an assignment that requires a file upload submission…

1. Log in as a student with google drive registered in their
   services and navigate to the assignments page.
2. Select an assignment that requires a file upload.
3. Click the "Submit Assignment" button.
4. Use keyboard shortcuts (tab and arrows) to tab from
   File Upload to Google Doc.
6. Navigate down the list using tab.
7. Expand a folder with space.
8. Select a file with space.

Change-Id: Ia8148472c2405c58cdfded760d7ed547edd1630e
Reviewed-on: https://gerrit.instructure.com/52450
Tested-by: Jenkins
QA-Review: Nathan Rogowski <nathan@instructure.com>
Reviewed-by: Matt Berns <mberns@instructure.com>
Product-Review: Strand McCutchen <smccutchen@instructure.com>