spec: fix a fragile wiki page spec, and mark it as pending
This spec was broken so that it usually produced a passing result. It is
testing that a cog icon regains focus after a dialog closes, but was not
waiting for ajax requests to complete before testing. If you wait for
ajax requests to complete, the cog icon loses the focus again, so the
functionality is broken.

So, the spec was fixed, and CNVS-20199 files to address the underlying
bug.

Change-Id: I85b7da78434ab3351a58d44cfbb95092e137705a
Reviewed-on: https://gerrit.instructure.com/53346
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Product-Review: Zach Wily <zach@instructure.com>
QA-Review: Zach Wily <zach@instructure.com>