graded fake student submissions get counted
the 'x / y graded' text in speegrader was not including graded
fake student (also known as Student View student) submissions
in the 'x' portion. the text now includes graded fake student
submissions.

closes CNVS-20090

test plan:

as a teacher, create an assignment. then go to the course settings
page and click 'Student View' in the upper right-hand corner (this
will cause you to masquerade as a fake student). complete the
assignment as a fake student and submit it. stop masquerading
as a fake student and log back in as a teacher. go to the speedgrader.
notice that the text in the upper right-hand corner of the speedgrader
shows '0 / X Graded', where X is the number of students in the course.
grade the Test Student's submission in speedgrader. verify that the text
in the upper right-hand corner of speedgrader changes to '1 / X Graded'.

Change-Id: Ib915c1409defcf8895d125820a0712d2262c075c
Reviewed-on: https://gerrit.instructure.com/53436
Tested-by: Jenkins
Reviewed-by: Derek Bender <dbender@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>