remove submitting form when hitting enter
fixes CNVS-20731
fixes CNVS-20730
fixes CNVS-20352

When adding a user at the account level KO, errors would appear right
when the dialog was open. This ensures that doesn't happen.

When you press tab -> enter in the edit name dialog for an ePortfolio
and creating a course, it should not automatically submit the form.

This re-adds the fixDialogButtons to modules dialog so buttons don't
loose their style. See this gerrit.
https://gerrit.instructure.com/#/c/54892/2

Test Plan

Given I'm working in the account
And I go to the users tab
And I'm only using the keyboard to navigate
When I focus on the "add user" button
And I press "enter"
Then I the new user dialog should appear
And that dialog should NOT have any inital errors for the form fields

When you first log into canvas, click on the "create course" button
Using japanse characters, enter something into the course name
Tab to the character you want and press enter
The form should not submit
The characters should be entered into the text field

Create an ePortfolio with a title
Edit the title by click on "ePortfolio Settings" (a dialog should
appear)
Use Hiragana to enter text into the text field. Once the chinese
characters appear, tab to one of them, hit enter.
The characters should be inserted into the text field and it should not
close the dialog window and submit the form. Pressing enter the second
time should submit the form.

Change-Id: I8ea187e09406258b146a645d7d3ceecdb1e297e8
Reviewed-on: https://gerrit.instructure.com/55165
Tested-by: Jenkins
QA-Review: Clare Strong <clare@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>