Fix styles for Post Grades button in gradebook toolbar
fixes: CNVS-20288

The Post Grades button was moved to the gradebook
toolbar in SIS-768.

It is sliding behind the other buttons on the right-hand
side of the toolbar. The single button also is rendered
inside an unordered list when it doesn't need to be.
http://cl.ly/image/2D2L1F3L172b

QA Test Plan:
- Open the Gradebook.
- Make sure the Post Grades button is enabled and
  showing in the toolbar above the main
  gradebook content.
- Make sure the Post Grades button is not sliding
  behind the other buttons (see image link above).
  It should now look like this:
  http://cl.ly/image/3v0l3u1o3o2t
- Verify Post Grades button is still working:
  https://guides.instructure.com/m/4152/l/295587-how-do-i-publish-grades-to-powerschool

Change-Id: I04bab4869ecddeef054426c18921ef8e58b8167e
Reviewed-on: https://gerrit.instructure.com/53591
Tested-by: Jenkins
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
QA-Review: Ben Bolton <bbolton@instructure.com>
Product-Review: Chris Hart <chart@instructure.com>