fix issues with aux build failing
fixes CNVS-20885

Test that were passing models into filesytemObjectThumbnail were
failing because of a timeout that sometimes gets reached. This fixes
the problem by setting the 'thumbnail_url' property on File objects
which will ensure that .fetch never gets called even if the timeout is
reached. A better solution is to ensure that all tests that call
anything that works it way down to filesytemObjectThumbnail have a
fakeClock that ticks and never makes it past the timeout.

Change-Id: I198685aeed8174fcd565ce42a6a0d5bbe109340b
Reviewed-on: https://gerrit.instructure.com/55690
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>
QA-Review: Sterling Cobb <sterling@instructure.com>