Create thumnail link for LTI launch inside the RCE
When we get a ConentItemSelection response that supports a
thumbnail we need to create an HTML element with the information
necessary to allow a plugin to create the actual overlay.

This is only for LtiLink Content Items, with a presentation document
target of "iframe" and a thumbnail or icon attribute

Ref: PLAT-1082

Test Plan:
use the test tool to create an lti link with a thumbnail
make sure it shows up with a thumbnail.
PLAT-1083 will civer testing the actual functionality

Change-Id: Id90f11e26571a61a153d5a825670ef8e1636d38e
Reviewed-on: https://gerrit.instructure.com/56343
Tested-by: Jenkins
Reviewed-by: Nathan Mills <nathanm@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Brad Horrocks <bhorrocks@instructure.com>