Create LtiThumbnailLauncher
This plugin will load on every page. It adds a delegate click handler
for 'lti-thumbnail-launch' to the body. The handler will replace the
element clicked with an iframe to the launch URL, it uses the target
attribute to store the placement info.

Ref: PLAT-1083

Test Plan:
Create different lti launches with a mixture of the following options
target: [window, iframe]
thumbnail: [{
      "@id": "http://developers.imsglobal.org/images/imscertifiedsm.png",
      "width" : 147,
      "height" : 184
    }, null]

This can easily be accomplished with the lti test tool

Change-Id: I408fac500b6a38de0ea81ef79e8c0e46e105b4a7
Reviewed-on: https://gerrit.instructure.com/56344
Tested-by: Jenkins
Reviewed-by: Nathan Mills <nathanm@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Brad Horrocks <bhorrocks@instructure.com>