make course copy date fields accessible
closes CNVS-19320
refs CNVS-15543

This applies a screenreader-only label to the datetime pickers
used in the course copy UI, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

TEST PLAN:
- go to copy a course
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the course copy with no
  other context.
- tab through to the same fields and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: Iaaca5e83db3a940bf1a0271a2e0803e5afcd169e
Reviewed-on: https://gerrit.instructure.com/56515
Tested-by: Jenkins
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>