Let users determine if alpha statistic is valid
Show alpha statistic as long as we can compute it, it's up to the
teacher if the statistic is valid or useful. As such the new requirment
for alpha data is 2 or more questions and variance > 0, variance > 0
implies 2 or more submissions with different scores.

Fixes CNVS-20337

Test plan
 - test with 1 question
     alpha = N/A
 - test with 2 questions and 1 submission
     alpha = N/A
 - test with 2 questions and 2 submissions with same score
     alpha = N/A
 - test with 2 or more questions and 2 or more submissions with
 different scores
     alpha = a number

Change-Id: I8b1e91b24f75c9f0311863b8573a03b2b8b3d9dc
Reviewed-on: https://gerrit.instructure.com/56580
Reviewed-by: Ryan Taylor <rtaylor@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: Jason Sparks <jsparks@instructure.com>
Tested-by: Jenkins