make access token exipration accessible
closes CNVS-19328
refs CNVS-15543

This applies a screenreader-only label to the datetime picker
used in the access tokens UI, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

ALSO: adds a new override option to tooltips to skip
collision detection in instances when you always want the
tooltip to show up on a certain side no matter what.

ALSO AGAIN: removes the convention of using "flipfit flipfit"
for the default collision detection and uses just "flipfit" instead.
Per jQuery UI docs, position: {collision: ""}
accepts either one or two values.  If one
it uses the same value for horizontal or vertical, if two it uses
them individually (one for horizontal one for vertical). Thus it's
redundant to provide two values if they're the same.

ALSO FINALLY: adds a spec file for the tooltip.coffee file,
which has not had them before

TEST PLAN:
1) go to your profile page
2) bring up the dialog to create an access token
3) Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the expiration date with no
  other context.
4) tab through to the same field and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: I3591711acbd7a3311e0dfa0969f2a3d5a511bb6b
Reviewed-on: https://gerrit.instructure.com/56637
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>