make course section editing accessible
closes CNVS-19330
refs CNVS-15543

This applies a screenreader-only label to the datetime pickers
used in the section editing UI, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

TEST PLAN:
- go to edit a course section (/courses/x/sections/y)
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the start and end dates with no
  other context.
- tab through to the same fields and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: Iea4156379b914e3cc2f386e6921e3e7b2336c43d
Reviewed-on: https://gerrit.instructure.com/56649
Tested-by: Jenkins
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>