Check for user logged in before fetching with filesEnv
This makes it so that a user needs to be logged in for the
filesEnv.rootFolders to map properly.  Because we were fetching
within, when that fetch failed it was causing an error to be
thrown on the modules page.

The only time that this fails is when a user is not logged in
and the files nav link is disabled (in a public course)

fixes CNVS-21197

Test Plan:
  - Go to a course that is publicly available
  - Add files to some modules
  - Disable the files link in the course navigation
  - Go incognito (make sure you put on the proper disguise to do so
    https://goo.gl/V2OXnZ)
  - You should be able to load up the modules page of your course
    without getting an error.
  - * Bonus Points * Make sure new files is still working okay.
    I guess we should just call it files now since it's on for
    everyone and there is only one true files now.
    So just make sure that files isn't borked beyond all belief.

Change-Id: Id24b052c2b041091eb3ab46fe2723ed815b32af3
Reviewed-on: https://gerrit.instructure.com/56814
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>