round gradebook summary column even if group weights are enabled
fixes CNVS-21172

somewhat related to CNVS-21106

Rounds grade totals in the gradebook to 2 decimal places when the option
"Set Group Weights" > "Weight final grade based on groups" is enabled.
It previously rounded correctly when this setting was disabled; the
behavior when this setting was enabled was overlooked

Note that trailing zeros will not be shown. For example, a score of
exactly 73.3% will be shown as "73.3%" rather than "73.30%"

Test plan:
* Go into the gradebook, click on the options gear, and enable the
  checkbox under "Set Group Weights" > "Weight final grade based on
  groups"
* Ensure that the total grades are rounded to 2 decimal places
* In general, make sure all grade displays in the gradebook are rounded
  to 2 decimal places

Change-Id: I91600cf1daa86e4c5a8cb4b6a1fe7a05d4ceef49
Reviewed-on: https://gerrit.instructure.com/57022
Tested-by: Jenkins
Reviewed-by: Cameron Matheson <cameron@instructure.com>
QA-Review: Derek Hansen <dhansen@instructure.com>
Product-Review: Christi Wruck