Theme Editor: take out custom placeholder styling
closes: CNVS-21899

In the theme editor, when you set, for example,
"brand-primary" to red, a bunch of other inputs
inherit that same color as their default value.
to show that the value is inherited, we show the
red color as the color picker swatch and we also
set the placeholder of the input to "red"

it was confusing because with this custom placeholder
styling, you couldn't tell if you had explicitly
set the second value or if it was just like that
because you had set the brand-primary value
that it inherits from.

when I git blamed these lines, they were from my
big commit so I don't know if it was something
that I accidentally added or if it was intended

Change-Id: I2b6dc2c53903b6471b07e78529563da9a0be6275
Reviewed-on: https://gerrit.instructure.com/58313
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>
Tested-by: Jenkins
QA-Review: Nathan Rogowski <nathan@instructure.com>