Add Login to Theme Editor
This adds the login branding to our Theme Editor. You can
now customize the Login screen!

closes: CNVS-19361
closes: CNVS-22446

Test Plan:
- Turn on New UI
- Go to Theme Editor (Account > Settings > Open Theme Editor)
- Go to Login Screen
- Edit all the things!
- Apply the styles
- Logout > Go to login screen

Here is a layout of what can be customized:
http://cl.ly/image/1m043S0x0e1w

Note:

- Currently we don't have a good way of previewing applying the styles. You
need to apply the styles fully in order to preview it.
There is another ticket to address this issue.

- This does not (and should not) affect Multi-Auth or Sign up screens. These
should still remain the same white/light color scheme.

- This also sets the default background image on login to be nothing
which fixes CNVS-22446

- We've taken out the ability to use svg for the login logo. There are some
complexities with loading these in as a background and properly applying the
right sizes and will be tackled in a future ticket

Change-Id: I1185cbdf41ffab082a7cf1d51c1e0c403dbb8188
Reviewed-on: https://gerrit.instructure.com/58479
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Tested-by: Jenkins