fix gb assignment dropdown menue event removal
when the gradebook is rerendered via slickgrid
the events in the column header that are attached
to the menu dropdown are removed. to fix this
we call @fixColumnReordering(), which reattaches
the events. this bug occured because when we
change the editability of an assignment, slickgrid
re renders. the menu also didn't show up after reordering
the secondary id / student name columns.

fixes CNVS-21599

test plan
  - login as a teacher into a gradebook with
    multiple grading periods
  - click a dropdown in an assignment column header
  - ensure the dropdown works

Change-Id: I8df6c6bf55dcf91fd78c7e59aff49035b117284f
Reviewed-on: https://gerrit.instructure.com/58515
Tested-by: Jenkins
Reviewed-by: Cody Poll <cpoll@instructure.com>
QA-Review: Derek Hansen <dhansen@instructure.com>
Product-Review: Dylan Ross <dross@instructure.com>