fix js specs to elimnate orphaned "save changes?" dialogs
in our js specs, long after the wikiPageEditView
was removed from the page, we'd still fire the
"do you want to save your changes before you leave"
on beforeunload of the window.

this fixes it so we only show that dialog if 
the element rendered by the backbone view is still
in the page.

test plan: 
* go to wiki page,
* make changes, don't save
* try to navigate away from the page
* you should still see the message asking you:
  are you sure? you'll lose changes!

Change-Id: I646cc80e8dc24c186e81b6396fcf1c034143d246
Reviewed-on: https://gerrit.instructure.com/58583
Reviewed-by: James Williams  <jamesw@instructure.com>
Tested-by: Jenkins
Product-Review: James Williams  <jamesw@instructure.com>
QA-Review: James Williams  <jamesw@instructure.com>