fix issue with inserting links inside tables
fixes CNVS-21903

When in tinymce, if you had a table with a link inside of it and you
selected the link, then selected something from the wikisidebar it
would remove all of the styles and surrounding html from the link. This
commit ensures that the link changes but html and styles continue to be
preserved.

Test Plan

Open up tinymce where there is a wikisidebar (pages for instance)
Create a table in the text editor
Insert a link into a table row/column
Add surrounding <span> or <div> tags to the link via html editing tool
Use the wikisidebar to modify the link by selecting it with your cursor
and then clicking on a link in the wikisidebar.
Check the html in tinymce. Ensure that the span/div and all styles are
still there.

Change-Id: I290f0fc24e61475f50492ffbde35d5f720b48c4f
Reviewed-on: https://gerrit.instructure.com/59196
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Charles Kimball <ckimball@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>