New UI: show the "courses" menu item all the time
closes: CNVS-22368

When a student only has enrollments in unpublished
courses, they do not see a courses global nav link.
The link is needed to get to "all courses" and see
their enrollments.

this changes things so the "courses" menu is always
shown. that way those users will be able to get
to that "all courses" link

Product / UX/UI review question:
This is just one possible way of fixing this problem.
the other thing we could do is to leave it how it was
but do a second, more expensive DB query to see if
there are any of those unpublished or archived courses that
I can see on the /courses page but not in the course
menu. in standup, we agreed to do it like this but
if any of you feel strongly that we should do it
the other way, that's fine too.

test plan:
Add a new user
Enroll them in an unpublished course
Masquerade as the new user,
click the courses global nav option
you should see the "all courses" link in the tray popup

Change-Id: I140dbc4f4d2852c386cc57cac3f2afec6ef62979
Reviewed-on: https://gerrit.instructure.com/60233
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>