OAuth and Confirmation Box re-style
closes: CNVS-20724

This commit brings the new login styles to Canvas'
many login confirmation screens.

QA: Please note that this commit does not touch the
User Terms page or the OAuth2 request permissions
page. Those will be addressed in subsequent tickets.

Test Plan:

There are a lot of pages to this, and Chart has
documented the things that we know and are aware
of that these changes touch here:

https://docs.google.com/document/d/1csN1YJOm9zcl-bzUaA5yT3deUWlENleTyYuN8_VGhFE/

- See document
- Go through the document, and make sure all the
  pages work as advertised :)

Change-Id: I2b9d85643a01e436613e22761a71286b3bcc07a0
Reviewed-on: https://gerrit.instructure.com/61561
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Myller de Araujo <myller@instructure.com>